# docs/api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "FUTABUS HTTP API",
    "version": "1.0.0",
    "description": "Current public HTTP API. Booking and account pages use web routes and are not represented as REST endpoints. The SePay webhook is called by SePay, not by customers."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current FUTABUS host"
    }
  ],
...
  "paths": {
...
    "/api/v1/tokens": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "operationId": "createToken",
        "summary": "Issue a 30-day API token",
        "description": "Only active users with verified email can obtain a token. Rate-limited to 10 requests per minute.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password",
                  "device_name"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "format": "password"
                  },
                  "device_name": {
                    "type": "string",
                    "maxLength": 100
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Token issued",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "token_type",
                    "access_token",
                    "expires_at"
                  ],
                  "properties": {
                    "token_type": {
                      "type": "string",
                      "const": "Bearer"
                    },
                    "access_token": {
                      "type": "string"
                    },
                    "expires_at": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation or credential error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "SePayHmac": {
        "type": "apiKey",
        "in": "header",
        "name": "X-SePay-Signature",
        "description": "Requires X-SePay-Timestamp too. The signature covers the raw body and timestamp; a static key entered in Swagger UI cannot generate it."
      },
      "SePayApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Exact value: Apikey <configured webhook key>. Used only when API-key mode is configured."
      },
      "BearerToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Sanctum personal access token"
      }
    },
    "schemas": {
...
      "ValidationError": {
        "type": "object",
        "required": [
          "message",
          "errors"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
    }
  }
}